+++ CHANGELOG.md
@@ -132,7 +132,10 @@
 - `color` and `without-content-background` props to ipl-expanding-space
 - Added `copiable` to ipl-data-row
 - ipl-data-row can contain any element as the value through the default slot
-- Allow removing the background color on ipl-button
+- Updated ipl-button:
+  - Allow removing the background color
+  - Added `inline` prop
+  - Button size can be adjusted by changing `font-size` in CSS
 
 # 2.10.1